fix: Fix handle for binary operator click
Fix binary operator click handle for case where the last
character is `-` and second last character is either a
closing bracket or a left unary operator. In which case
the click of the binary operator will have no effect on
the expression.
For example:
expression = 5+(-
binary operator clicked = +
final expression = 5+(-

Signed-off-by: Archit Raj <arch10@ymail.com>